requirements.txt
aiogram==3.25.0
aiohttp==3.13.3
python-dotenv==1.2.1
playwright==1.58.0
